print invoice

+++ database/migrations/2023_08_16_184835_create_patient_sessions_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->foreign('patient_id')->references('id')->on('patients');
             $table->string('patient_type')->default('OUTPATIENT');
             $table->string('consultation_type')->default('GENERAL');
-            $table->integer('registration_fee')->nullable();
-            $table->integer('consultation_fee')->nullable();
+            $table->decimal('registration_fee', 10, 2)->nullable();
+            $table->decimal('consultation_fee', 10, 2)->nullable();
             $table->unsignedBigInteger('doctor_id')->nullable();
             $table->foreign('doctor_id')->references('id')->on('users');
             $table->dateTime('discharged')->nullable();
